Fixed 64-Bit Clang build
At least on Windows. Had to do a bunch of changes, because Visual Studio
is apparently a non-standards compliant piece of shit. Looks like some
unit tests are broken and need to be fixed. Not sure if it's new issues
or just shit that never would've actually worked outside VS.